Create global functions manager
